requirements.txt
-f /usr/share/pip-wheels
alembic==0.9.6
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-Login==0.4.0
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
mysql-connector-python==2.0.4
ordereddict==1.1
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python-editor==1.0.3
python-slugify==1.2.4
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.1.15
traitlets==4.3.2
Unidecode==0.4.21
wcwidth==0.1.7
Werkzeug==0.12.2
WTForms==2.1
